+++ CHANGELOG.md
@@ -9,6 +9,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ### Added
 
+- Query parameter `allusers` in `GET /compute/jobs` to show all visible jobs for the user in the scheduler
 - Environment variable `UVICORN_LOG_CONFIG` to enable [Uvicorn log configuration](https://www.uvicorn.org/settings/#logging) file path (analog to `--log-config`)
 
 ### Changed
@@ -18,6 +19,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Fixed reservation start and end datatime parsing.
 - Handles instances where no Job exit status is provided.
 - Fixed unnecessary user keys retrieval with SSH connection pool.
+- Fixed proper SSH process termination on timeout.
 
 ## [2.2.6]